requirements/doc.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
accessible-pygments==0.0.4
    # via pydata-sphinx-theme
alabaster==0.7.13
    # via sphinx
asgiref==3.7.2
    # via
    #   -r requirements/test.txt
    #   django
babel==2.12.1
    # via
    #   pydata-sphinx-theme
    #   sphinx
beautifulsoup4==4.12.2
    # via pydata-sphinx-theme
build==1.0.3
    # via -r requirements/doc.in
certifi==2023.7.22
    # via
    #   -r requirements/test.txt
    #   requests
cffi==1.16.0
    # via
    #   -r requirements/test.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.3.0
    # via
    #   -r requirements/test.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/test.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.5.0
    # via
    #   -r requirements/test.txt
    #   edx-toggles
coverage[toml]==7.3.1
    # via
    #   -r requirements/test.txt
    #   pytest-cov
cryptography==41.0.4
    # via
    #   -r requirements/test.txt
    #   pyjwt
    #   secretstorage
django==3.2.21
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
    #   django-crum
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
django-crum==0.7.9
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
    #   edx-toggles
django-environ==0.11.2
    # via -r requirements/test.txt
django-waffle==4.0.0
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
djangorestframework==3.14.0
    # via
    #   -r requirements/test.txt
    #   drf-jwt
    #   edx-drf-extensions
doc8==1.1.1
    # via -r requirements/doc.in
docutils==0.19
    # via
    #   doc8
    #   pydata-sphinx-theme
    #   readme-renderer
    #   restructuredtext-lint
    #   sphinx
drf-jwt==1.19.2
    # via
    #   -r requirements/test.txt
    #   edx-drf-extensions
edx-django-utils==5.7.0
    # via
    #   -r requirements/test.txt
    #   edx-drf-extensions
    #   edx-toggles
edx-drf-extensions==8.10.0
    # via -r requirements/test.txt
edx-opaque-keys[django]==2.5.1
    # via
    #   -r requirements/test.txt
    #   edx-drf-extensions
edx-toggles==5.1.0
    # via -r requirements/test.txt
exceptiongroup==1.1.3
    # via
    #   -r requirements/test.txt
    #   pytest
idna==3.4
    # via
    #   -r requirements/test.txt
    #   requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==6.8.0
    # via
    #   build
    #   keyring
    #   sphinx
    #   twine
importlib-resources==6.1.0
    # via keyring
iniconfig==2.0.0
    # via
    #   -r requirements/test.txt
    #   pytest
jaraco-classes==3.3.0
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via
    #   -r requirements/test.txt
    #   code-annotations
    #   sphinx
keyring==24.2.0
    # via twine
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   -r requirements/test.txt
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.1.0
    # via jaraco-classes
newrelic==9.1.0
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
nh3==0.2.14
    # via readme-renderer
packaging==23.1
    # via
    #   -r requirements/test.txt
    #   build
    #   pydata-sphinx-theme
    #   pytest
    #   sphinx
pbr==5.11.1
    # via
    #   -r requirements/test.txt
    #   stevedore
pkginfo==1.9.6
    # via twine
pluggy==1.3.0
    # via
    #   -r requirements/test.txt
    #   pytest
psutil==5.9.5
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/test.txt
    #   cffi
pydata-sphinx-theme==0.14.1
    # via sphinx-book-theme
pygments==2.16.1
    # via
    #   accessible-pygments
    #   doc8
    #   pydata-sphinx-theme
    #   readme-renderer
    #   rich
    #   sphinx
pyjwt[crypto]==2.8.0
    # via
    #   -r requirements/test.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==3.13.0
    # via
    #   -r requirements/test.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/test.txt
    #   edx-django-utils
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via
    #   -r requirements/test.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.txt
pytest-django==4.5.2
    # via -r requirements/test.txt
python-slugify==8.0.1
    # via
    #   -r requirements/test.txt
    #   code-annotations
pytz==2023.3.post1
    # via
    #   -r requirements/test.txt
    #   babel
    #   django
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   -r requirements/test.txt
    #   code-annotations
readme-renderer==42.0
    # via twine
requests==2.31.0
    # via
    #   -r requirements/test.txt
    #   edx-drf-extensions
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==1.0.0
    # via twine
restructuredtext-lint==1.4.0
    # via doc8
rfc3986==2.0.0
    # via twine
rich==13.6.0
    # via twine
secretstorage==3.3.3
    # via keyring
semantic-version==2.10.0
    # via
    #   -r requirements/test.txt
    #   edx-drf-extensions
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==6.2.1
    # via
    #   -r requirements/doc.in
    #   pydata-sphinx-theme
    #   sphinx-book-theme
sphinx-book-theme==1.0.1
    # via -r requirements/doc.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.4
    # via
    #   -r requirements/test.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/test.txt
    #   code-annotations
    #   doc8
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via
    #   -r requirements/test.txt
    #   python-slugify
tomli==2.0.1
    # via
    #   -r requirements/test.txt
    #   build
    #   coverage
    #   doc8
    #   pyproject-hooks
    #   pytest
twine==4.0.2
    # via -r requirements/doc.in
typing-extensions==4.8.0
    # via
    #   -r requirements/test.txt
    #   asgiref
    #   edx-opaque-keys
    #   pydata-sphinx-theme
    #   rich
urllib3==2.0.5
    # via
    #   -r requirements/test.txt
    #   requests
    #   twine
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources
